fix: webpack v5 support
With webpack v5, this module does not seem to work properly. The problem happens
when external assets (like images) are required by the template. The resulting
assets seem to be corrupted (at least for .png's).

This could be due to the fact that the `file-loader` was replaced by
`asset/resource` in webpack v5.

Also, this module contains a "fallback" `file-loader.js` file which is used when
there are no loaders for this type of asset. However, it looks like it's now
broken in webpack v5, and the fact that there's this fallback loader prevents
`asset/resource` from requiring the asset properly.

To fix this, we simply remove this fallback loader from this module. It's never
actually used in practice anyway.